requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.25.0
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Vector operations and similarity search
faiss-cpu>=1.7.0
scikit-learn>=1.0.0

# LangChain and NLP
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Image processing and computer vision
opencv-python>=4.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.62.0
click>=8.0.0
python-dotenv>=0.19.0

# Jupyter and development
jupyter>=1.0.0
ipywidgets>=7.6.0
notebook>=6.4.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Optional: Web interface
streamlit>=1.28.0
gradio>=3.0.0

# Optional: API integration
openai>=0.27.0
huggingface-hub>=0.16.0

# Data handling
jsonlines>=3.0.0
PyYAML>=6.0

# Logging and monitoring
loguru>=0.6.0

# Memory optimization
psutil>=5.8.0
